refactor: 프로젝트 상세페이지 훅 분리
